refactor: simplify Filter types and fix breaking changes
