+++ frontend/src/components/jobs/search/search-bar.tsx
@@ -3,7 +3,6 @@ import { Input } from "@mantine/core";
 import { IconSearch } from "@tabler/icons-react";
 
 export default function SearchBar() {
-
   return (
     <Input
       placeholder="Search for a company or a role..."
