+++ frontend/src/context/jobs/filter-provider.tsx
@@ -6,19 +6,19 @@ import { FilterContext, FilterState, initialState } from "./filter-context";
 import { useRouter } from "next/navigation";
 import { CreateQueryString } from "@/lib/utils";
 
-export function FilterProvider({
-  children,
-}: {
-  children: ReactNode;
-}) {
+export function FilterProvider({ children }: { children: ReactNode }) {
   const [filters, setFilters] = useState<FilterState>(initialState);
   const router = useRouter();
 
   const updateFilters = (newFilters: Partial<FilterState>) => {
-    setFilters(curr => ({ ...curr, ...newFilters }));
+    setFilters((curr) => ({ ...curr, ...newFilters }));
     const params = CreateQueryString(newFilters);
     router.push(`/jobs?${params}`);
   };
 
-  return <FilterContext.Provider value={{filters, updateFilters}}>{children}</FilterContext.Provider>;
+  return (
+    <FilterContext.Provider value={{ filters, updateFilters }}>
+      {children}
+    </FilterContext.Provider>
+  );
 }
