+++ frontend/src/types/job.ts
@@ -17,25 +17,3 @@ export interface Job {
   createdAt: string;
   updatedAt: string;
 }
-
-// frontend/src/types/filters.ts
-export interface JobFilters {
-  search: string;
-  studyFields: string[];
-  jobTypes: string[];
-  locations: string[];
-  workingRights: string[];
-  page: number;
-  sortBy: "recent" | "relevant";
-}
-
-// frontend/src/types/api.ts
-export interface JobsResponse {
-  jobs: Job[];
-  total: number;
-}
-
-export interface ApiError {
-  message: string;
-  code: string;
-}
